+++ src/lib/components/NoticeCard.svelte
