+++ src/lib/components/ProgressBar.svelte
