+++ src/lib/interfaces.ts
