+++ src/lib/prompt.ts
